% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_avg_2021.R
\docType{data}
\name{batting_avg_2021}
\alias{batting_avg_2021}
\title{Batting Averages 2021}
\format{
231 obs. of 5 variables:
\describe{
\item{Player}{Name of player}
\item{lgID}{League}
\item{H}{Hits}
\item{AB}{At bats}
\item{AVG}{Batting average}
}
}
\usage{
batting_avg_2021
}
\description{
Batting Averages 2021
}
\keyword{datasets}
