% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_players.R
\docType{data}
\name{four_players}
\alias{four_players}
\title{Four Players Home Plate Statistics}
\format{
64 obs. of 12 variables:
\describe{
\item{PX}{chr: ?}
\item{PZ}{chr: ?}
\item{BIP}{?}
\item{H}{?}
\item{HR}{?}
\item{H_Rate}{?}
\item{HR_Rate}{?}
\item{Z_H}{?}
\item{Z_HR}{?}
\item{Player}{chr: Player name}
\item{px}{?}
}
}
\usage{
four_players
}
\description{
Four Players Home Plate Statistics
}
\keyword{datasets}
