\name{Rxdata-package}
\alias{Rxdata-package}
\docType{package}
\title{
Data for the Book "R by Example".
}
\description{
Description: Data for examples and exercises in the books "R by Example" by Jim Albert and Maria Rizzo.
}
\author{
Jim Albert and Maria Rizzo
}
\references{
Jim Albert and Maria Rizzo (2012). \emph{R by Example}, Springer.

Jim Albert and Maria Rizzo (2024). \emph{R by Example, 2nd ed.}, Springer (in preparation).
}
\keyword{ package }
\keyword{ data }
