% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.record.mile.R
\docType{data}
\name{world.record.mile}
\alias{world.record.mile}
\title{World Record Mile Data}
\format{
276 obs. of  3 variables:
\describe{
\item{Gender}{female or male}
\item{Time}{chr: time as "mm:ss"}
\item{X}{The whole minutes part of Time}
\item{X.1}{The "ss" part of Time}
\item{seconds}{time expressed in seconds}
\item{Athlete}{Name}
\item{Nationality}{nationality}
\item{Date}{chr: date}
\item{Year}{year}
}
}
\source{
Wikipedia page \url{https://en.wikipedia.org/wiki/Mile_run_world_record_progression}
}
\usage{
world.record.mile
}
\description{
Mile run world record progression  as recorded by the
International Amateur Athletics Federation (IAAF).  The
dataset includes 32 world records for men ratified by
the IAAF, and 29 world records for women both in the
pre-IAAF and IAAF eras.
}
\keyword{datasets}
