\name{Rxdata-package}
\alias{Rxdata-package}
\docType{package}
\title{
Data for the Book "R by Example".
}
\description{
Data for examples and exercises in the book "R by Example" by Jim Albert and Maria Rizzo.
}
\author{
Jim Albert and Maria Rizzo
}
\references{
Jim Albert and Maria Rizzo (2012). \emph{R by Example}, Springer.
}
\keyword{ package }
\keyword{ data }
